% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findNNs.R
\name{findNNs}
\alias{findNNs}
\title{Read a file and compute the NN through permutations;}
\usage{
findNNs(filename, cell_names = NULL)
}
\arguments{
\item{filename}{the file to be processed}

\item{cell_names}{A list of cell types}
}
\value{
The Nearest Neibour Cell distance in the file
}
\description{
Read a file and compute the NN through permutations;
}
