% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFileCount.r
\name{getFileCount}
\alias{getFileCount}
\title{Get the count number of each phenotype for each tissue category}
\usage{
getFileCount(dataFile)
}
\arguments{
\item{dataFile}{File path to the data file}
}
\value{
a data frame summary of count
}
\description{
Get the count number of each phenotype for each tissue category
}
