% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/procDist.R
\name{procDist}
\alias{procDist}
\title{Extract the statistics from the result given by findNNs}
\usage{
procDist(distResult, thred = 30)
}
\arguments{
\item{distResult}{A distance result file from findNNs}

\item{thred}{a threshold distance as cut-off}
}
\value{
The statistics extracted from the distace result file
}
\description{
Extract the statistics from the result given by findNNs
}
\examples{
Cell.X.Position=sample(1:1000, 20, replace=TRUE)
Cell.Y.Position=sample(1:1000, 20, replace=TRUE)
distance = runif(20,min=0,max=100)
distResult = data.frame (Cell.X.Position,Cell.Y.Position,distance)
thred=30
procDist(distResult,thred)
}
